Fix yet another bug in UV mapping logic for spheres
I was linearly mapping the normal's y coordinate to [0, 1] instead of
using spherical coordinates.

A shameful mistake tbh. If it turns out that there are even more bugs
in the UV mapping logic for spheres, I'll stop working on the project
and try something more aligned with my math skills (like gardening).